+++ lib/slim_lint/linter/consecutive_control_statements.rb
@@ -6,12 +6,19 @@ module SlimLint
 
     on [:multi] do |sexp|
       Utils.for_consecutive_items(sexp,
-                                  ->(nested_sexp) { nested_sexp.match?([:slim, :control]) },
+                                  method(:flat_control_statement?),
                                   config['max_consecutive'] + 1) do |group|
         report_lint(group.first,
                     "#{group.count} consecutive control statements can be " \
                     'merged into a single `ruby:` filter')
       end
+    end
+
+    private
+
+    def flat_control_statement?(sexp)
+      sexp.match?([:slim, :control]) &&
+        sexp[3] == [:multi, [:newline]]
     end
   end
 end